Add files via upload
Group Project completed with fellow teammates for the MSIS 2613_DBMS class. The focus of the final report is to simulate the database system of a company through database creation, database modeling, and discussion. Our group chose to center the project on Walmart. The final report contains: Walmart's business objectives; a list of entities and relationships for relational modeling; images of ERD and EER models (EER is in 3rd normal form); code for SQL database creation; sample queries to simulate questions about Walmart's database, and more.